pyproject.toml
[project]
name = "openstreetmap-images"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.1",
    "pillow>=12.1.1",
    "requests>=2.32.5",
]
